Remove warnings

+++ libraries/render-utils/src/RenderDeferredTask.cpp
@@ -104,12 +104,12 @@ RenderDeferredTask::RenderDeferredTask() : Task() {
     
     // Lighting Buffer ready for tone mapping
     _jobs.push_back(Job(new ToneMappingDeferred::JobModel("ToneMapping")));
-    _toneMappingJobIndex = _jobs.size() - 1;
+    _toneMappingJobIndex = (int)_jobs.size() - 1;
 
     // Debugging Deferred buffer job
     _jobs.push_back(Job(new DebugDeferredBuffer::JobModel("DebugDeferredBuffer")));
     _jobs.back().setEnabled(false);
-    _drawDebugDeferredBufferIndex = _jobs.size() - 1;
+    _drawDebugDeferredBufferIndex = (int)_jobs.size() - 1;
 
     // Status icon rendering job
     {
@@ -118,7 +118,7 @@ RenderDeferredTask::RenderDeferredTask() : Task() {
         auto statusIconMap = DependencyManager::get<TextureCache>()->getImageTexture(iconMapPath);
         _jobs.push_back(Job(new render::DrawStatus::JobModel("DrawStatus", renderedOpaques, DrawStatus(statusIconMap))));
         _jobs.back().setEnabled(false);
-        _drawStatusJobIndex = _jobs.size() - 1;
+        _drawStatusJobIndex = (int)_jobs.size() - 1;
     }
 
     _jobs.push_back(Job(new DrawOverlay3D::JobModel("DrawOverlay3D")));
